qt a11y: Destroy QWindow when popup gets hidden
When a popup gets hidden, destroy its associated
QWindow.

This prevents a "top-level" a11y object being reported
when a popup gets shown, then hidden again, e.g. after
expanding the Paragraph Style combobox in Writer's
Formatting toolbar, then closing it again.

Showing the popup results in a QWindow being created
that would not implicitly be destroyed when the
popup gets hidden again.

Due to popups in LO currently not using the Qt::Popup
type, but Qt::ToolTip, they are considered for top-level
a11y children of the app, see the commit message of

    Change-Id: I7aff5c435dfa8b6aadcbbedb0d84db19bb86c8ab
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Sep 19 09:15:37 2024 +0200

        qt a11y: Defer QWindow creation until frame gets shown

for more details.

Only delete the QWindow for non-spontaneous hide events,
as the QWidget::hideEvent doc [1] says:

> Note: A widget receives spontaneous show and hide events when its
> mapping status is changed by the window system, e.g. a spontaneous hide
> event when the user minimizes the window, and a spontaneous show event
> when the window is restored again. After receiving a spontaneous hide
> event, a widget is still considered visible in the sense of isVisible().

With this commit in place, after starting Writer, opening the
Paragraph Style combobox, then closing it again, only the
Writer window is shown as a top-level child of the app
in Accerciser when using the qt6 VCL plugin, as expected.

Showing the popup again works just fine, as a new QWindow
is implicitly created then.

[1] https://doc.qt.io/qt-6/qwidget.html#hideEvent

Change-Id: Iefa5d05ea128966c4417d53d122a6a0f1178fc00
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/173657
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins