i#94666 sw: layout: fix problem with WIDOW_MAGIC in sections
This was already fixed in CWS sw301bf03 with commit
5559afee02fc2be18cded35a17a03aa8191b08f5 but then broken again, perhaps
by commit f2e3655255db4032738849cd4b77ce67a6e2c984 "Avoid
-fsanitize=signed-integer-overflow", which changed a magic constant,
effectively disabling the fix.

The problem (in a different document than attached at the bug) is that
the first text frame 128 in a section frame 258 gets its height set to
WIDOW_MAGIC in CalcPreps(), which grows the section frame to the maximum
allowed by its upper, and then when the real size of the text frame is
set it shrinks the section frame to be far too small, so the last text
frames and the whole table remain formatted at a position on the page
but are not painted because the paint is cut off at the (wrong) bottom
of the section frame.

(On master, the problem with the internal document cannot be reproduced
due to some other change which causes the text frame at the cut-off
position to have mbFramePrintAreaValid=false which causes it to MoveFwd
and that calls SwSectionFrame::SimpleFormat() which fixes the height,
but that all looks accidental.)

Change-Id: If13d993a0cab5701f45223a70b2c5c8b0690ebeb