further winget configuration files to complete the setup
this is still a chicken-and-egg situation, since the steps are meant to
run before the repository is cloned, however it makes some sense to
have that in the repo since deps and similar are tied to the code after
all.

The user-side makes use of all script resources, which of course is not
ideal, but still better than nothing.

Unfortunately the admin side uses hardcoded paths - can be switched to
also using script resources instead of the xRemoteFile DSC, but that is
trading one bad thing for another…

Future tweaks might include setting up a Dev Drive and/or whitelisting
directories to be excluded from AV-checks

Change-Id: Ide1f0efb94264a6c72a9e87c2c212aa527d7d330
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/173676
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>
Tested-by: Jenkins