cool#9992 lok doc sign: extract duplicated code to SfxLokHelper
I want to improve extractCertificate() so it can work on a certificate
chain, but that's easier when this code is not directly in desktop/, but
at some lower level.

This allows covering the code with tests from CppunitTest_sfx2_view in a
follow-up change.

If this code will be needed by some non-LOK area as well, then it can be
moved down further, but let's wait for a second area first.

(cherry picked from commit 7bad76c3a5b319f8ede74be8f78e5645f9ffd050)

Change-Id: I6291da0c3e56aed7dca1a8dc1446209044cace92
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/173811
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>