tdf#130857 Refactor VclBuilder::handlePacking/applyPackingProperties
Instead of doing XML parsing in
`VclBuilder::applyPackingProperty`, call
`BuilderBase::collectProperty` in
`VclBuilder::handlePacking` to get a string map
of the properties and pass that as a param
to the renamed `VclBuilder::applyPackingProperties`
that can then work with the key-value pairs right
away.

This is one step to split XML parsing out of methods
that directly work with vcl::Window, in order to
be able to reuse the logic for other implementations.

(`git show --ignore-space-change` shows the
"actual" change more clearly.)

Change-Id: If29cbd7a13da2732c46e4a0b0b50d0963525e729
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/173714
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>