tdf#130857 VclBuilder: Avoid using vcl WindowType to detect toolbar
Instead of checking that the type of the vcl::Window
is WindowType::TOOLBOX, rely on the class name of the
current item to be one of the toolbar item classes
in `VclBuilder::handleObject` and instead double-check
and make the code in `VclBuilder::applyAtkProperties`
conditional on this actually being a `ToolBox` instead
of asserting.

(Not sure whether there's a valid case for using toolbar
items outside of toolbars, but avoid potential regressions
here if that's used anywhere.)

This gets rid of one vcl::Window specific way of handling
things in `VclBuilder::handleObject`, which is one step
for preparing it for reuse by other builders in the future
in particular QtBuilder from pending WIP Gerrit change [1].

In a quick test, the accessible name of
"Show Up to Outline Level" is still properly set
for the corresponding toolbar item in the navigator only
and not for the toolbar, see

    commit ad1167b92b8b8fdabf6b21e945682c0bafc80946
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Feb 27 13:12:43 2024 +0100

        tdf#159910 a11y VclBuilder: Apply tool item's a11y name to itself

for background.

[1] https://gerrit.libreoffice.org/c/core/+/161831

Change-Id: I8a7dbbee9b88bcb7c0cd6bc862d2c5cc4a290af4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/173734
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>