tdf#163064: pic element is required here, after all
In commit cf15306ccf49da290b391517e2c5dd22a4f1be45 (ERROR: Invalid
content was found starting with element 'p:pic'., 2014-12-22), the
pic element inside the oleObj element was only alloswed for the old
revision of ECMA-736 export, because that was considered invalid by
officeotron. However, as of ECMA-376-1:2016,  this element is
mandatory; CT_OleObject definition in Annex A has:

 <xsd:element name="pic" type="CT_Picture" minOccurs="1" maxOccurs="1"/>

and Annex L (Primer) has L.7.2.5 "Embeddings in a PresentationML
Document" saying:

 The oleObj element shall have a pic child element that (optionally)
 contains the image data to be used in place of loading the actual
 object data.

The omission of this in the export is the reason of tdf#163064. So
here I filter out the error from the validation results, which is
not ideal (I have no way to know if the found pic is really in the
oleObj, or somewhere else), but a lesser evil, compared to required
exclusion of all tests that export OLE objects.

Change-Id: Ia73a49da7347e8ff22c626e211b55ba1e0625070
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/173761
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins