tdf#163086 Resize SVG pin icons to 24x24 and clean them up
This also reverts commit 13a42d5c2d433da6c2c69159bfc6df0e37643333 which
worked around the issue of oversized SVGs.

Sukapura Dark's pin_document.svg was already 24x24, gets a cleanup anyway.

Editing the SVG source, I changed the 64 (and in one case 512) to 24.

Then, I opened each one in Inkscape, clicked the lock icon for width and height,
input /2.6667 (or 21,3333 in case the viewbox used to be 512) to the end of the
existing width and hit Enter to apply. Then in Object - Align and Distribute
panel, I set Relative to: Page and centered on vertical and horizontal axes.

Finally, I used svgcleaner to clean out all the useless crap in the source.

Change-Id: Ic42bd5fce3227f008fe98eed304b1d565cf13651
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/173774
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>
Tested-by: Jenkins
Reviewed-by: Ilmari Lauhakangas <ilmari.lauhakangas@libreoffice.org>