tdf#163086 Resize SVG pin icons to 24x24 and clean them up
Sukapura Dark's pin_document.svg was already 24x24, gets a cleanup anyway.

Editing the SVG source, I changed the 64 (and in one case 512) to 24.

Then, I opened each one in Inkscape, clicked the lock icon for width and height,
input /2.6667 (or 21,3333 in case the viewbox used to be 512) to the end of the
existing width and hit Enter to apply. Then in Object - Align and Distribute
panel, I set Relative to: Page and centered on vertical and horizontal axes.

Finally, I used svgcleaner to clean out all the useless crap in the source.

Change-Id: Id5608de8dd735ab8efeee4d2417be07a3109f6d0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/173783
Reviewed-by: Ilmari Lauhakangas <ilmari.lauhakangas@libreoffice.org>
Tested-by: Jenkins
(cherry picked from commit 65773351521b0a3c84d34915e5b0537610effc40)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/173789