cool#9992 lok doc sign: store signing cert in the view
In case signing certificate would be stored in the model (like it
happens on the desktop), then the certificate viewer would show all
signing certs of the currently editing users, which is far from
expected.

At the same time, there is just one NSS database per process, so it's
hard to not have them in the same database.

Fix the problem by storing a reference to the signing key in the view
shell, then we can show only that in xmlsecurity/ in LOK mode, and thing
else.

This is just the sfx2/ side of things, the xmlsecurity/ part still needs
implementing.

Change-Id: Id7a56f28598ec60217c6bac095c2ff350f39bc3b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/173837
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins