tdf#163091: check rParam.mnX before passing it to RowInfo::cellInfo
When moving to the left, the RowInfo may have e.g. cells 21 to 41,
and rParam.mnX may be 15, which is outside of the RowInfo data.

It seems that in this case, when we handle the data outside of
the visible area, setting clip marks may be safely skipped anyway.

Change-Id: I2f507dc166468c44e34645bc34c827456f1ab1ec
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/173857
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
(cherry picked from commit 8b23abac4929bbb15a52afb083259cd2d24b4995)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/173802
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>