tdf#162268 sw: ApplyParagraphMarkFormatToEmptyLineAtEndOfParagraph
Commit 2b47fae7e3e23ee7c733708500cb0482ad7f8af1 introduced the
compatibility setting ApplyTextAttrToEmptyLineAtEndOfParagraph, but that
was probably a mistake.

What Word is doing there is not applying a text attribute but applying the
formatting of the paragraph marker; add a new compatibility setting
ApplyParagraphMarkFormatToEmptyLineAtEndOfParagraph to do this.

Change SwAttrIter to apply the RES_PARATR_LIST_AUTOFMT formatting when
the position behind the last character is reached, and use it to set the
height of the last line in SwLineLayout::CalcLine() in case it is empty
or contains only spaces/tabs.

Frustratingly this requires another change to fdo74110.docx to get rid
of some odd font that's applied to the paragraph marker.

Also, change SwTextFrame::IsHiddenNow() to take into account paragraph
marker formatting; if all characters are hidden but the paragraph marker
isn't hidden, the paragraph is still displayed in Word.

Change-Id: Icccd3e822ad0301ccbe373b50431c3254f691d6e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/173880
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
(cherry picked from commit 7b8cec8322b60673e686660caf875bc71117bbd0)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174065
Reviewed-by: Thorsten Behrens <thorsten.behrens@allotropia.de>