SwNavigator: Improve the page change control
The intent of this patch is to address an issue with the Writer
Navigator page change control reported at https://
ask.libreoffice.org/t/writer-display-single-page/111050

"I have selected View > Normal and View > Whitespace and View> Zoom>
Entire Page. On the bottom right, I’ve chosen the single page icon."
... "It is odd that the page number displayed as you click the
[Navigator Page control spin button] up and down arrows does not
change.."

The patch changes the function used by the page control to determine
the page number of the page in the view to that used by the status
bar page control. It also disables page tracking by the control when
the view is set to Multiple-page view or Book view.

Change-Id: I439bb4493061528f87eaeca3c08a33095adb8192
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/173888
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>