tdf#130857: Move .ui -> VCL respone code mapping to handleActionWidget
Move mapping the (GTK) response code set in the .ui file
to the VCL equivalent from VclBuilder::set_response
to BuilderBase::handleActionWidget.

As VCL response codes are used by all implementations, this
is not specific to the vcl::Window implementation and
also allows an upcoming QtBuilder::set_response_code
implementation to use the passed value as is.

Change-Id: I5b6d262a657932038c423c0c2cdcf1fd32c8b6be
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/173930
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins