fix(lok): Expose the character encoding dialog
The character encoding dialog appears when importing a dialog with an
ambiguous character encoding. At the point when it appears, the document
isn't loaded, so we have to handle it in a similar way to the text
import dialog or password prompt dialog, allowing it to function early.

It's also required to make the dialog into a jsdialog, as tunneled
dialogs do not work this early in the document loading process.

I have not asynced the dialog here because there's no point. I did write
code where I set up a dialog factory to do this similar to how other
modules (sw, sc, sd, etc.) handle dialogs and make them async.
Unfortunately, this dialog is not useful with multiple users, since as
it's used on opening a document. It may, in the future, be useful to
someone to make a patch that asyncs the other writerperfect dialog (epub
export) so I have uploaded and abandoned my dialog factory patch as
Id8ec474510f3c8637639ad59331156c6e75dafb4.

Documents that require this dialog, whether async or not, cause crashes
when someone opens the document before the first person has finished
responding to this dialog. I still consider this behavior better than
crashing whenever we attempt to show this dialog at all.

Change-Id: Ibd36e695e733e07053bb6e22a510ac988be3ded2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174044
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>
(cherry picked from commit 283f771b92a14c06871c931b01174622b855abe5)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174950
Tested-by: Jenkins