qt: Just ignore request to set frame to invalid screen number
Instead of setting the screen to the primary screen,
just ignore the request when QtFrame::SetScreenNumber
gets called with an invalid screen number.

The previous handling was introduced in

    commit f806a2832aee62efc0e0404f7c24d53aaaf814d0
    Date:   Mon Oct 22 16:35:21 2018 +0200

        tdf#120451: Use primary screen if requested screen doesn't exist

to fix a crash, but just ignoring the request
avoids crashing, too.

(`git show --ignore-space-change` helps to see
the "actual" change more easily.)

Change-Id: I272ec64f89ef835161fa68c7fdc8cd0e407b7495
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174033
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins