qt: No longer use SalFrame::maGeometry in QtFrame at all
With previous commit

    Change-Id: Icaf712c956297dbc7a774bbd995d42eb8a63a1bb
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Sep 27 11:06:55 2024 +0200

        qt: Query screen on demand in QtFrame

in place, all relevant members are explicitly set
for the returned SalFrameGeometry in
QtFrame::GetUnmirroredGeometry, so there's no more
reason to start with a copy of `maGeometry` instead of
a default-constructed SalFrameGeometry.

Margins for decorations are never set for the Qt-based
VCL plugins (using methods like SalFrameGeometry::setLeftDecoration),
so the default of 0 is fine for all of these.

Change-Id: Ibed98c665e61a4ba06bf9087a159048417598798
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174035
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins