tdf#162268 sw: IgnoreHiddenCharsForLineCalculation
The other problem here is that if hidden text isn't displayed,
SwLineLayout::CalcLine() doesn't ignore the hidden text portion when
calculating the line height.

This is apparently the case all the way back to OOo 3.3, but it doesn't
make much sense, so change the default to ignore hidden text here,
except if a pre-existing ODF document is imported.

Change-Id: Ief3f62bbdb6ac0ac79b54d6885a711128ad8d81f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174041
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>