tdf#130857 qt weld: Set widget layout and add to parent layout
When creating a QLayout, pass the parent widget
to the ctor as well, which makes the layout be
set for the widget. Quoting from the QVBoxLayout ctor doc [1]

> Constructs a new top-level vertical box with parent parent.
>
> The layout is set directly as the top-level layout for parent.
> There can be only one top-level layout for a widget. It is returned
> by QWidget::layout().

Also add widgets to their parent layout if the parent
is a QLayout, and not a QWidget.

This makes the items of the "Help" -> "License" dialog
actually be part of the dialog in a WIP branch
that has "QtInstanceBuilder::IsUIFileSupported" in
the set of supported files in QtInstanceBuilder::IsUIFileSupported.

[1] https://doc.qt.io/qt-6/qvboxlayout.html#QVBoxLayout-1

Change-Id: Ib823896a28520b6f7e1e89244239d3204d1d10e4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174072
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>