tdf#130857 qt weld: Ensure dialog button box is last in layout
For QDialog, make sure that if a button box is included in
the dialog's layout, that this is the last item in the layout,
by removing from the layout and adding it at the end again.

I don't see any explicit child index explicitly being set in
the .ui file for the "Help" -> "License Information" dialog
("sfx/ui/licensedialog.ui").

Potentially GTK implicitly visually makes the dialog's
button box the last item in the dialog.

Corresponding child is this one:

        <child internal-child="action_area">
          <object class="GtkButtonBox" id="dialog-action_area1">

Potentially the `<child internal-child="action_area">`
identifies this as the dialog's button box that should
be last, at least dialog newly created in glade
also has that set.

For QMessageBox, which is a QDialog subclass, this
special handling is not needed, as its default
button box is used, which is already at the right
place.

This addresses the first aspect mentioned in
previous commit:

    Change-Id: Ic9393755ec474f77ff22a1115e3cccba9d7b26cb
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Sat Sep 28 00:07:28 2024 +0200

        tdf#130857 qt weld: Add initial support for dialog and label

> However, currently buttons and the label with the text
> are in the wrong order (i.e. buttons are above the text)

Still missing:

> and clicking the buttons doesn't yet have any effect.

Change-Id: Id991551548c1e54fdf2e169886a6c67fc307931f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174077
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins