tdf#130857 qt weld: Close dialog on button click
For QDialogs (created for "GtkDialog" objects in .ui files)
that have a button box, let clicking any of the buttons
that have a response code set close the dialog with
that response code by connecting to the
QAbstractButton::clicked signal.

This addresses the remaining aspect for the
"Help" -> "License Information" dialog
mentioned in earlier commit

    Change-Id: Ic9393755ec474f77ff22a1115e3cccba9d7b26cb
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Sat Sep 28 00:07:28 2024 +0200

        tdf#130857 qt weld: Add initial support for dialog and label

> However, currently (...) clicking the buttons doesn't yet have any
> effect.

Actually using Qt(InstanceBuilder) for that dialog
will be enabled in an upcoming commit.

Change-Id: Ie3c26963318fd5832688e3ed73f6952e09da4bb7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174079
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins