tdf#130857 qt a11y weld: Set GtkBuilder ID as widget's accessible ID
Set the GtkBuilder ID specified in .ui files as the accessible
ID for the QWidget created for the object, in line with what
happens for other implementations as well. This is basically
the equivalent of

    commit c64b823b74cbd3063d6e87643bd68ea5343b2f54
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Jul 16 13:57:41 2024 +0200

        tdf#155447 qt a11y: Report accessible ID

for the case where native Qt widgets are used instead of
vcl::Window.

Other than QWidget, QLayout is not part of the a11y
hierarchy of an application, so there's nothing to
do for these.

This e.g. makes Accerciser show an accessible ID of
"LicenseDialog" for the "Help" -> "License Information"
dialog.

The accessible ID set this way currently isn't reported
on the AT-SPI level for QMessageBox (e.g. the "Save Document?"
dialog) however. Suggested upstream change to make that
work as well: [1]

[1] https://codereview.qt-project.org/c/qt/qtbase/+/594390

Change-Id: I50e8ca6f8bea628df5a80d101c4a0a7bfded9c19
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174158
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>