wina11y: Unify handling for editable/readonly state a bit
The LABEL, STATIC and NOTIFICATION roles are
used for readonly text, so if the EDITABLE state were
ever set on these, that should be fixed in the underlying
XAccessibleContext implementation instead of special
handling to always set the readonly state in the
Windows a11y bridge. Therefore, rely on the
EDITABLE state not being set for these in order to
report the IAccessible/MSAA STATE_SYSTEM_READONLY
like for the other roles handled here.

Change-Id: Ie8398ad0186d57ebf92701ab3a2c0603b50f5a2e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174277
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins