wina11y: Bridge focusable state as is
Drop all the special-casing of particular (parent) roles
etc. when considering whether or not to set the focusable
state.
Instead, fully rely on the AccessibleStateType::FOCUSABLE
state as reported by the underlying XAccessibleContext
and report the IAccessible/MSAA STATE_SYSTEM_FOCUSABLE
state whenever the former is set.

I tend to think that if this this change results in
unexpected behavior, it is likely that issues should best
be fixed in the underlying XAccessibleContext
implementations rather than worked around in the Windows
a11y bridge.

Change-Id: I6e2eee3a9cbe779de3b3d349f23efc6a68a8344b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174278
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>