cool#9992 lok doc sign: update sign status after modify the list of trusted CAs
Load a document, sign it, "green" icon on the status bar. Reload the
document, turns into a "yellow" icon saying the CA is not trusted, when
it was already trusted before.

The trouble is that the document signature status is calculated on load,
and the CA to be trusted is only given later, as part of the
initialization of the LOK view.

Fix the problem by invalidating the signature state when a new CA is
trusted.

The test document was produced by signing an empty document using the
keys from xmlsecurity/qa/xmlsec/data/, which gives us a way to create a
signature that is initially not trusted.

Change-Id: I1e1dbf616ce54c4823d62104f838342de6870f52
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174307
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>